+++ cmd/thanos/store.go
@@ -36,10 +36,13 @@ func registerStore(m map[string]setupFunc, app *kingpin.Application, name string
 	chunkPoolSize := cmd.Flag("chunk-pool-size", "Maximum size of concurrently allocatable bytes for chunks.").
 		Default("2GB").Bytes()
 
-	objStoreConfig := regCommonObjStoreFlags(cmd, "")
+	objStoreConfig := regCommonObjStoreFlags(cmd, "", true)
 
 	syncInterval := cmd.Flag("sync-block-duration", "Repeat interval for syncing the blocks between local and remote view.").
 		Default("3m").Duration()
+
+	blockSyncConcurrency := cmd.Flag("block-sync-concurrency", "Number of goroutines to use when syncing blocks from object storage.").
+		Default("20").Int()
 
 	m[name] = func(g *run.Group, logger log.Logger, reg *prometheus.Registry, tracer opentracing.Tracer, debugLogging bool) error {
 		peer, err := newPeerFn(logger, reg, false, "", false)
@@ -63,6 +66,7 @@ func registerStore(m map[string]setupFunc, app *kingpin.Application, name string
 			name,
 			debugLogging,
 			*syncInterval,
+			*blockSyncConcurrency,
 		)
 	}
 }
@@ -86,14 +90,15 @@ func runStore(
 	component string,
 	verbose bool,
 	syncInterval time.Duration,
+	blockSyncConcurrency int,
 ) error {
 	{
-		bucketConfig, err := objStoreConfig.Content()
+		confContentYaml, err := objStoreConfig.Content()
 		if err != nil {
 			return err
 		}
 
-		bkt, err := client.NewBucket(logger, bucketConfig, reg, component)
+		bkt, err := client.NewBucket(logger, confContentYaml, reg, component)
 		if err != nil {
 			return errors.Wrap(err, "create bucket client")
 		}
@@ -113,6 +118,7 @@ func runStore(
 			indexCacheSizeBytes,
 			chunkPoolSizeBytes,
 			verbose,
+			blockSyncConcurrency,
 		)
 		if err != nil {
 			return errors.Wrap(err, "create object storage store")