+++ pkg/testutil/prometheus.go
@@ -9,17 +9,19 @@ import (
 	"os/exec"
 	"path/filepath"
 	"runtime"
+	"strings"
 	"syscall"
+	"testing"
 	"time"
 
-	"github.com/improbable-eng/thanos/pkg/block/metadata"
-
 	"github.com/go-kit/kit/log"
+	"github.com/improbable-eng/thanos/pkg/block/metadata"
 	"github.com/improbable-eng/thanos/pkg/runutil"
 	"github.com/oklog/ulid"
 	"github.com/pkg/errors"
 	"github.com/prometheus/tsdb"
 	"github.com/prometheus/tsdb/labels"
+	"github.com/prometheus/tsdb/testutil"
 	"golang.org/x/sync/errgroup"
 )
 
@@ -28,17 +30,18 @@ const (
 	defaultAlertmanagerVersion = "v0.15.2"
 	defaultMinioVersion        = "RELEASE.2018-10-06T00-15-16Z"
 
-	promBinEnvVar         = "THANOS_TEST_PROMETHEUS_PATH"
+	// Space delimited list of versions.
+	promVersionsEnvVar    = "THANOS_TEST_PROMETHEUS_VERSIONS"
 	alertmanagerBinEnvVar = "THANOS_TEST_ALERTMANAGER_PATH"
 	minioBinEnvVar        = "THANOS_TEST_MINIO_PATH"
 )
 
 func PrometheusBinary() string {
-	b := os.Getenv(promBinEnvVar)
-	if b == "" {
-		return fmt.Sprintf("prometheus-%s", defaultPrometheusVersion)
-	}
-	return b
+	return prometheusBin(defaultPrometheusVersion)
+}
+
+func prometheusBin(version string) string {
+	return fmt.Sprintf("prometheus-%s", version)
 }
 
 func AlertmanagerBinary() string {
@@ -60,9 +63,10 @@ func MinioBinary() string {
 // Prometheus represents a test instance for integration testing.
 // It can be populated with data before being started.
 type Prometheus struct {
-	dir    string
-	db     *tsdb.DB
-	prefix string
+	dir     string
+	db      *tsdb.DB
+	prefix  string
+	version string
 
 	running bool
 	cmd     *exec.Cmd
@@ -80,13 +84,40 @@ func NewTSDB() (*tsdb.DB, error) {
 	})
 }
 
+func ForeachPrometheus(t *testing.T, testFn func(t testing.TB, p *Prometheus)) {
+	vers := os.Getenv(promVersionsEnvVar)
+	if vers == "" {
+		vers = defaultPrometheusVersion
+	}
+
+	for _, ver := range strings.Split(vers, " ") {
+		if ok := t.Run(ver, func(t *testing.T) {
+			p, err := newPrometheus(ver, "")
+			testutil.Ok(t, err)
+
+			testFn(t, p)
+		}); !ok {
+			return
+		}
+	}
+}
+
 // NewPrometheus creates a new test Prometheus instance that will listen on local address.
+// DEPRECARED: Use ForeachPrometheus instead.
 func NewPrometheus() (*Prometheus, error) {
-	return NewPrometheusOnPath("")
+	return newPrometheus("", "")
 }
 
 // NewPrometheus creates a new test Prometheus instance that will listen on local address and given prefix path.
 func NewPrometheusOnPath(prefix string) (*Prometheus, error) {
+	return newPrometheus("", prefix)
+}
+
+func newPrometheus(version string, prefix string) (*Prometheus, error) {
+	if version == "" {
+		version = defaultPrometheusVersion
+	}
+
 	db, err := NewTSDB()
 	if err != nil {
 		return nil, err
@@ -99,10 +130,11 @@ func NewPrometheusOnPath(prefix string) (*Prometheus, error) {
 	}
 
 	return &Prometheus{
-		dir:    db.Dir(),
-		db:     db,
-		prefix: prefix,
-		addr:   "<prometheus-not-started>",
+		dir:     db.Dir(),
+		db:      db,
+		prefix:  prefix,
+		version: version,
+		addr:    "<prometheus-not-started>",
 	}, nil
 }
 
@@ -121,10 +153,11 @@ func (p *Prometheus) Start() error {
 
 	p.addr = fmt.Sprintf("localhost:%d", port)
 	p.cmd = exec.Command(
-		PrometheusBinary(),
+		prometheusBin(p.version),
 		"--storage.tsdb.path="+p.db.Dir(),
 		"--web.listen-address="+p.addr,
 		"--web.route-prefix="+p.prefix,
+		"--web.enable-admin-api",
 		"--config.file="+filepath.Join(p.db.Dir(), "prometheus.yml"),
 	)
 	go func() {
@@ -138,12 +171,17 @@ func (p *Prometheus) Start() error {
 	return nil
 }
 
-// Addr gets correct address after Start method.
+// Dir returns TSDB dir.
+func (p *Prometheus) Dir() string {
+	return p.dir
+}
+
+// Addr returns correct address after Start method.
 func (p *Prometheus) Addr() string {
 	return p.addr + p.prefix
 }
 
-// SetConfig updates the contents of the config file.
+// SetConfig updates the contents of the config file. By default it is empty.
 func (p *Prometheus) SetConfig(s string) (err error) {
 	f, err := os.Create(filepath.Join(p.dir, "prometheus.yml"))
 	if err != nil {
@@ -188,6 +226,30 @@ func CreateBlock(
 	mint, maxt int64,
 	extLset labels.Labels,
 	resolution int64,
+) (id ulid.ULID, err error) {
+	return createBlock(dir, series, numSamples, mint, maxt, extLset, resolution, false)
+}
+
+// CreateBlockWithTombstone is same as CreateBlock but leaves tombstones which mimics the Prometheus local block.
+func CreateBlockWithTombstone(
+	dir string,
+	series []labels.Labels,
+	numSamples int,
+	mint, maxt int64,
+	extLset labels.Labels,
+	resolution int64,
+) (id ulid.ULID, err error) {
+	return createBlock(dir, series, numSamples, mint, maxt, extLset, resolution, true)
+}
+
+func createBlock(
+	dir string,
+	series []labels.Labels,
+	numSamples int,
+	mint, maxt int64,
+	extLset labels.Labels,
+	resolution int64,
+	tombstones bool,
 ) (id ulid.ULID, err error) {
 	h, err := tsdb.NewHead(nil, nil, nil, 10000000000)
 	if err != nil {
@@ -252,8 +314,10 @@ func CreateBlock(
 		return id, errors.Wrap(err, "finalize block")
 	}
 
-	if err = os.Remove(filepath.Join(dir, id.String(), "tombstones")); err != nil {
-		return id, errors.Wrap(err, "remove tombstones")
+	if !tombstones {
+		if err = os.Remove(filepath.Join(dir, id.String(), "tombstones")); err != nil {
+			return id, errors.Wrap(err, "remove tombstones")
+		}
 	}
 
 	return id, nil